Update .gitignore to exclude bootstrap/cache/routes-v7.php and add logging facade to AuthController; create config.php for application settings

+++ backend/app/Http/Controllers/TaskController.php
@@ -131,7 +131,7 @@ class TaskController extends Controller
                 'title' => $request->title,
                 'description' => $request->description,
                 'assigned_to' => (int) $request->assigned_to,
-                'created_by' => Auth::id(),
+                'created_by' => (int) Auth::id(),
                 'due_date' => $request->due_date,
                 'priority' => $request->priority ?? 'medium',
                 'status' => $request->status ?? 'pending',
